FFMPEG auto install
FFMPEG is auto installed on startup if not present
error handling for functions dealing with files
